+++ .gitignore
@@ -1,7 +1,6 @@
 data/**/*.mat
 **/*.png
 **/*.pyc
-**/**/*.pyc
 **/*.jpeg
 **/SUNRGBDv2Test
 data/princeton_sunrgbd/SUNRGBD
