+++ utilities/train.py
@@ -1,10 +1,7 @@
-import os
-import numpy as np
 import math
+import tensorflow as tf
+import numpy as np
 import matplotlib.pyplot as plt
-from resnet_2d import *
-from data.cs231n.data_utils import get_CIFAR10_data
-from data.princeton_sunrgbd.load_data import *
 
 # Train the Model
 def train_model(device, sess, model, X_data, labels, epochs=1, batch_size=64,
@@ -73,46 +70,3 @@ def train_model(device, sess, model, X_data, labels, epochs=1, batch_size=64,
       plt.show()
 
   return total_loss, accuracy
-
-# Save Checkpoint of Model
-def save_model_checkpoint(session, saver, filename):
-  save_path = saver.save(session, filename)
-  print("Model checkpoint saved in file: %s" % save_path)
-
-# Recover Saved Model Checkpoint
-def recover_model_checkpoint(session, saver, filename):
-  saver.restore(session, filename)
-  print("Model restored!")
-
-def main():
-  # Suppress Annoying TensorFlow Logs
-  os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2'
-
-  # Test with CIFAR-10 Data
-  data = get_CIFAR10_data(num_training=49000, num_validation=1000, num_test=1000,
-                          subtract_mean=True)
-
-  # Create Model
-  print("Setting up model...")
-  model = setup_model()
-
-  # Create Session
-  sess = tf.Session()
-  sess.run(tf.global_variables_initializer())
-
-  # Define Device
-  device = '/gpu:0'
-
-  # Train Model
-  print("Training model...")
-  train_model(device, sess, model, data['X_train'], data['y_train'], epochs=1, batch_size=128,
-              is_training=True, log_freq=100, plot_loss=False)
-  print("Final Training Accuracy:")
-  train_model(device, sess, model, data['X_train'], data['y_train'], epochs=1, batch_size=64,
-              is_training=False, log_freq=100, plot_loss=False)
-  print('Final Validation Accuracy:')
-  train_model(device, sess, model, data['X_val'], data['y_val'], epochs=1, batch_size=64,
-              is_training=False, log_freq=100, plot_loss=False)
-
-main()
-exit(0)
